feat: Add China optimized version to Vercount
A China optimized version is now available for Vercount. This version is specifically designed for faster and more efficient access in China. Instructions on how to switch to the China optimized version are now included in the README file. This update also provides scripts for both the China optimized version and the global CDN version for quick use.