feat: enhance logging for data updates and retrievals
Improve logging for data updates and retrievals in the source code. This includes adding more detailed log messages for the update of page, site, and unique visitor data, as well as for the retrieval of this data before updates. Also, minor changes were made to the logging of the route handler in the application's log route.