feat: change logging level from info to debug
This commit changes all logger.info instances to logger.debug across multiple files in the application. This change is to ensure that only important information is displayed when the application is running in a production environment. It also includes a check in the logger.js file to only log debug messages when the application is not running in a production environment.